+++ bramble-core/src/main/java/org/briarproject/bramble/transport/agreement/TransportKeyAgreementValidator.java
@@ -42,7 +42,7 @@ class TransportKeyAgreementValidator extends BdfMessageValidator {
 	@Override
 	protected BdfMessageContext validateMessage(Message m, Group g,
 			BdfList body) throws FormatException {
-		MessageType type = MessageType.fromValue(body.getLong(0).intValue());
+		MessageType type = MessageType.fromValue(body.getInt(0));
 		if (type == KEY) return validateKeyMessage(m.getTimestamp(), body);
 		else if (type == ACTIVATE) return validateActivateMessage(body);
 		else throw new AssertionError();
