+++ bramble-core/src/test/java/org/briarproject/bramble/plugin/tcp/LanTcpPluginTest.java
@@ -217,13 +217,13 @@ public class LanTcpPluginTest extends BrambleTestCase {
 		// The plugin should have bound a socket and stored the port number
 		BdfList descriptor = kal.getDescriptor();
 		assertEquals(3, descriptor.size());
-		assertEquals(TRANSPORT_ID_LAN, descriptor.getLong(0).longValue());
+		assertEquals(TRANSPORT_ID_LAN, descriptor.getInt(0).intValue());
 		byte[] address = descriptor.getRaw(1);
 		InetAddress addr = InetAddress.getByAddress(address);
 		assertTrue(addr instanceof Inet4Address);
 		assertFalse(addr.isLoopbackAddress());
 		assertTrue(addr.isLinkLocalAddress() || addr.isSiteLocalAddress());
-		int port = descriptor.getLong(2).intValue();
+		int port = descriptor.getInt(2);
 		assertTrue(port > 0 && port < 65536);
 		// The plugin should be listening on the port
 		InetSocketAddress socketAddr = new InetSocketAddress(addr, port);
