+++ briar-core/src/main/java/org/briarproject/briar/privategroup/invitation/GroupInvitationValidator.java
@@ -56,7 +56,7 @@ class GroupInvitationValidator extends BdfMessageValidator {
 	@Override
 	protected BdfMessageContext validateMessage(Message m, Group g,
 			BdfList body) throws FormatException {
-		MessageType type = MessageType.fromValue(body.getLong(0).intValue());
+		MessageType type = MessageType.fromValue(body.getInt(0));
 		switch (type) {
 			case INVITE:
 				return validateInviteMessage(m, body);
